requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-celery-results==2.2.0
kombu==5.2.2
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
python-slugify==5.0.2
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
